+++ ReleaseNotes/release_notes.md
@@ -1,5 +1,6 @@
-
 ## New features:
+
+   1. add Gudermannian function and its inverse 
 
 ## Backward incompatible: